Added additional parameters to environment.
Added j_max, learn_jerk, num_circles, offset and offset_wall to environment. Also renamed max_a and max_v to a_max and v_max.

Changed the import of algorithm configuration so that it is possible to import configurations that not containing all parameters. Now it is possible to import configurations when the algorithm parameters are changed.